backend/pyproject.toml
[project]
name = "biohackagent"
version = "0.1.0"
description = "BioHackAgent project"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.11,<3",
  "pydantic-settings>=2,<3",
  "pymongo>=4.11,<5",
  "beanie>=2,<3",
  "python-dotenv>=1.0.0",
  "llama-index>=0.13.0",
  "llama-index-core>=0.13.0",
  "llama-index-llms-google-genai>=0.3.0",
  "llama-index-embeddings-google-genai>=0.3.0",
  "langchain-google-genai>=2.1.9",
  "langchain-google-vertexai>=2.0.28",
  "anthropic[vertex]>=0.62.0",
  "langsmith>=0.4.13",
  "uvicorn>=0.35.0",
  "selenium>=4.35.0",
  "webdriver-manager>=4.0.2",
  "charset-normalizer>=3.4.2",
  "mcp[cli]>=1.13.0",
  "brotli>=1.1.0",
  "async-timeout>=5.0.1",
  "langchain-mcp-adapters>=0.1.9",
  "firecrawl-py>=3.4.0",
  "google-api-python-client>=2.179.0",
  "google-auth-oauthlib>=1.2.2",
  "google-auth-httplib2>=0.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]

